[FLINK-23064][connector-kafka] Expose connector options as PublicEvolving
This closes #16334.
